requirements.txt
python-dotenv
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.13.0
scikit-learn>=1.0.0
pytesseract>=0.3.0
Pillow>=9.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
mysql-connector-python>=8.0.0
sqlalchemy>=2.0.0
pyodbc>=5.0.0
dotenv
